Add example value for MEV minimum bid
The example value makes it easier to understand what format is expected.
Without the example it is not obvious if this should be set in ETH or GWei
without looking up the MEV boost documentation.

We should also advocate for setting this to 0.05 ETH to improve
censorship resistance of the network as recommended by Vitalik
in this Tweet: https://twitter.com/VitalikButerin/status/1595793106926014466