Siren: Fix connecting to BN/VC
For some reason the BN and VC urls resolve to ipv6 addresses in the
Siren container, which is not supported by those endpoints. We therefore
force Siren to prefer ipv4 using an env variable for NodeJS
applications.

Note that this does not seem to be related to IPV6=true/false in .env .