Siren fixes (#1978)
* Siren: Fix when CL_NODE is comma-separated list

Siren would use the BEACON_URL as-is to connect to, which works for
the Lighthouse VC but not for Siren. If it is a comma-separated list of
BNs, we just pick the first one to connect to.

* Siren: Fix connecting to BN/VC

For some reason the BN and VC urls resolve to ipv6 addresses in the
Siren container, which is not supported by those endpoints. We therefore
force Siren to prefer ipv4 using an env variable for NodeJS
applications.

Note that this does not seem to be related to IPV6=true/false in .env .

---------

Co-authored-by: Pietje Puk <pietje@pietjepuk.net>